Setup basic structure + basic db tables
Fairly simple structure: we have a repo and a service; handlers talk to
the service, the service deals with the business object layer, the
service talks to the repo which only handles the database layer.

Trying out faux, instead of mockall.

No frontend in place yet, but it'll be askama + htmx + json ext for htmx
-- I can't deal with writing form data requests. I just can't.